(reports/): Separa em classes diferentes, cada um dos reports, e chama em main.py
